full work
I uploaded the work folder with all the model scripts to GitHub. The system input datasets are in the data/input/timeseries folder. I created multiple scenarios in order to compare different system configurations.

I have a crucial question about my model. First of all, the model does not need marginal costs = epsilon micro-costs to be assigned to define the PV distribution in order to work, since PyPSA solves the optimization problem by minimizing the OPEX objective function, optimizing power flows based on input datasets, link efficiencies and capacities, and the sizing of the technologies included in the model.

The question is, since my goal is to begin the transition to a PED, and therefore maximize self-sufficiency, how to set it up. The fundamental focus is on how to dispatch the distribution of PV, energy that has no operating production cost. If the model were not given the option to export it, then the system would simply tend to distribute the PV, optimizing flows and minimizing costs, thereby maximizing self-sufficiency.

When we add the option to export PV, since the sales price dataset is the same as the purchase price dataset, the system tends to trade rather than maximize self-sufficiency: it tends to export electricity when the price is high and it is advantageous to do so, and to import it when the price is lower.

For this reason:
- I set the input data in the model corresponding to the actual dataset of electricity import prices and the price of thermal imports, but choosing the electricity export price equal to a decimal epsilon value, so that the solver has a small benefit from selling the PV surplus, without however prevailing over the optimization of the objective function (only real prices were used to calculate the annual costs incurred).

-The alternative would be to set a constraint that allows the export of an amount of energy less than or equal to the energy produced by the PV, minus the energy supplied to the building, minus the energy that can enter the battery, minus the energy that can enter the heat pump, minus the energy that can enter the DSM Electric Storage.

I look forward to your advice, which is essential for me to continue with my work.

Thank you.